Split CatalogPluginsKotlinDSLIntegrationTest
We split the largest unrolled method into a separate class so
that we can cut the time into half.
